requirements.txt
beautifulsoup4==4.5.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-modelcluster==2.0
django-taggit==0.18.3
django-toolbelt==0.0.1
django-treebeard==4.0.1
djangorestframework==3.5.1
gunicorn==19.6.0
html5lib==0.999999999
Pillow==3.4.2
psycopg2==2.6.2
pytz==2016.7
six==1.10.0
static3==0.7.0
Unidecode==0.4.19
wagtail==1.7
webencodings==0.5
Willow==0.4
